+++ lib/src/core/model/identity.g.dart
@@ -1,0 +1,28 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// ignore_for_file: implicit_dynamic_parameter, non_constant_identifier_names, type_annotate_public_apis, omit_local_variable_types, unnecessary_this
+
+part of 'identity.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+Identity _$IdentityFromJson(Map<String, dynamic> json) => Identity(
+      transient: json['transient'] as bool?,
+      identifier: json['identifier'] as String,
+    );
+
+Map<String, dynamic> _$IdentityToJson(Identity instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('transient', instance.transient);
+  val['identifier'] = instance.identifier;
+  return val;
+}