+++ lib/src/core/model/loading.dart
@@ -1,0 +1,1 @@
+enum FlagsmithLoading { loading, loaded }