+++ test/fg/flagsmith_inmemory_test.dart
@@ -11,7 +11,7 @@ void main() {
     late FlagsmithClient fs;
     setUp(() async {
       fs = await setupClientAdapter(StorageType.inMemory, caches: true);
-      setupAdapter(fs, cb: (config, _adapter) {});
+      setupAdapter(fs, cb: (config, adapter) {});
     });
     tearDown(() {
       fs.close();
@@ -47,10 +47,10 @@ void main() {
       expect(result, isNotNull);
       expect(result.length, seeds.length);
 
-      final value = fs.hasCachedFeatureFlag(notImplmentedFeature);
+      final value = fs.hasCachedFeatureFlag(notImplementedFeatureName);
       expect(value, false);
 
-      final value1 = await fs.hasFeatureFlag(notImplmentedFeature);
+      final value1 = await fs.hasFeatureFlag(notImplementedFeatureName);
       expect(value1, false);
     });
     test('When get Features then success', () async {
@@ -59,6 +59,11 @@ void main() {
     });
     test('When get Features for user then success', () async {
       var user = Identity(identifier: 'test_sample_user');
+      setupEmptyAdapter(fs, cb: (config, adapter) {
+        adapter.onPost(fs.config.identitiesURI, (server) {
+          return server.reply(200, jsonDecode(identitiesResponseData));
+        }, data: user.toJson());
+      });
       var result = await fs.getFeatureFlags(user: user);
       expect(result, isNotNull);
       expect(result, isNotEmpty);
@@ -74,7 +79,7 @@ void main() {
       expect(resultNext, isNotEmpty);
     });
     test('When flag is not presented then false', () async {
-      var result = await fs.isFeatureFlagEnabled(notImplmentedFeature);
+      var result = await fs.isFeatureFlagEnabled(notImplementedFeatureName);
       expect(result, false);
     });
     test('When flag is presented then true', () async {
@@ -83,7 +88,7 @@ void main() {
     });
 
     test('When flag is not presented then value is null', () async {
-      var result = await fs.getFeatureFlagValue(notImplmentedFeature);
+      var result = await fs.getFeatureFlagValue(notImplementedFeatureName);
       expect(result, isNull);
     });
 
@@ -94,23 +99,23 @@ void main() {
     });
 
     test('When feature flag remove then success', () async {
-      final _featureName = 'my_feature';
+      final featureName = 'my_feature';
       await fs.reset();
 
-      final _current = await fs.hasFeatureFlag(_featureName);
-      expect(_current, true);
+      final current = await fs.hasFeatureFlag(featureName);
+      expect(current, true);
 
-      var result = await fs.removeFeatureFlag(_featureName);
+      var result = await fs.removeFeatureFlag(featureName);
       expect(result, true);
 
-      final _removed = await fs.hasFeatureFlag(_featureName);
-      expect(_removed, false);
+      final removed = await fs.hasFeatureFlag(featureName);
+      expect(removed, false);
     });
   });
 
   group('[InMemory storage] flags api failures', () {
     late FlagsmithClient fs;
-    final _identity = Identity(identifier: 'invalid_users_another_user');
+    final identity = Identity(identifier: 'invalid_users_another_user');
     setUp(() async {
       fs = await setupClientAdapter(StorageType.inMemory, caches: true);
     });
@@ -118,8 +123,8 @@ void main() {
       fs.close();
     });
     test('When get flags then 404 error', () async {
-      setupEmptyAdapter(fs, cb: (config, _adapter) {
-        _adapter
+      setupEmptyAdapter(fs, cb: (config, adapter) {
+        adapter
           ..onGet(
             config.flagsURI,
             (server) => server.throws(
@@ -136,35 +141,35 @@ void main() {
                 requestOptions: RequestOptions(path: config.flagsURI),
               ),
             );
-          }, queryParameters: _identity.toJson());
+          }, queryParameters: identity.toJson());
       });
       expect(() => fs.getFeatureFlags(), throwsA(isA<FlagsmithApiException>()));
       expect(fs.cachedFlags, isNotEmpty);
-      expect(() => fs.getFeatureFlags(user: _identity),
+      expect(() => fs.getFeatureFlags(user: identity),
           throwsA(isA<FlagsmithApiException>()));
     });
     test(
         'When get flags returns statusCode > 200 && < 300, then success but empty',
         () async {
-      setupEmptyAdapter(fs, cb: (config, _adapter) {
-        _adapter
+      setupEmptyAdapter(fs, cb: (config, adapter) {
+        adapter
           ..onGet(config.flagsURI,
               (server) => server.reply(201, jsonDecode('''[]''')))
-          ..onGet(fs.config.identitiesURI, (server) {
+          ..onPost(fs.config.identitiesURI, (server) {
             return server.reply(201, null);
-          }, queryParameters: _identity.toJson());
+          }, data: identity.toJson());
       });
       expect(await fs.getFeatureFlags(), isEmpty);
-      expect(await fs.getFeatureFlags(user: _identity), isEmpty);
+      expect(await fs.getFeatureFlags(user: identity), isEmpty);
     });
 
     test('When get user flags returns null, then success but empty', () async {
-      setupEmptyAdapter(fs, cb: (config, _adapter) {
-        _adapter.onGet(fs.config.identitiesURI, (server) {
+      setupEmptyAdapter(fs, cb: (config, adapter) {
+        adapter.onPost(fs.config.identitiesURI, (server) {
           return server.reply(200, null);
-        }, queryParameters: _identity.toJson());
+        }, data: identity.toJson());
       });
-      expect(await fs.getFeatureFlags(user: _identity), isEmpty);
+      expect(await fs.getFeatureFlags(user: identity), isEmpty);
     });
   });
 }